rename: SeoulDataPipelineProperties config/properties 하위로 이동
